vrpn_Analog.C: cast to get rid of warning vrpn_Nidaq.C: got rid of internal Sleep(1) for nice behavior -- users should 	use a timeout in mainloop instead. vrpn_FileConnection.h: added comment to indicate that fileconnection is based 	on vrpn connection, not synchronized connection vrpn_Connection.h: changed declaration of check_connection to be compatible wiht 	new timeout for mainloop. vrpn_Connection.C: got rid of some warnings, have time stamps sync-adjusted 	before logging, added timeout to check_connection so mainloop will 	sleep for up to timeout whether or not it is connected

+++ vrpn_Analog.C
@@ -220,7 +220,7 @@ int vrpn_Analog_Remote::handle_change_message(void *userdata,
 	vrpn_ANALOGCHANGELIST *handler = me->change_list;
 
 	cp.msg_time = p.msg_time;
-	cp.num_channel = ntohd(params[0]);
+	cp.num_channel = (long)ntohd(params[0]);
 
 	vrpn_float64 * chandata = (vrpn_float64 *) params+1;
 	for (vrpn_int32 i=0; i< cp.num_channel; i++)
